Move install to own module
For clarity reasons
